% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treetable.R
\name{levelsToGroups}
\alias{levelsToGroups}
\title{Translates hierarchy levels into unique group names which represent the elements hierarchy position.}
\usage{
levelsToGroups(levels)
}
\arguments{
\item{levels}{Vector of hierarchy levels.}
}
\value{
Vector of group names.
}
\description{
Translates hierarchy levels into unique group names which represent the elements hierarchy position.
}
\examples{
levels <- c(1, 2, 2, 1, 2, 3, 3, 2, 1)
levelsToGroups(levels)
# c("1+", "11", "12", "2+", "21+", "211", "212", "22", "3")
}

