% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treetable.R
\name{treetable}
\alias{treetable}
\title{Title}
\usage{
treetable(data, levels)
}
\arguments{
\item{data}{data.frame containing data to be shown}

\item{levels}{hierarchy levels}
}
\value{
html widget to display treetable data. See datatable function of
package DT
}
\description{
Title
}
\examples{
data <- data.frame(a = 1:9, b = c(5, 2, 3, 9, 4, 3, 1, 5, 3))
levels <- c(1, 2, 2, 1, 2, 3, 3, 2, 1)
treetable(data, levels)
}

