% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treetable.R
\name{levelsToGroups}
\alias{levelsToGroups}
\title{Translates hierarchy levels into unique group names which represent the
elements hierarchy position. An '_exp' is added at the end of the groups name
if it has child elements.}
\usage{
levelsToGroups(levels)
}
\arguments{
\item{levels}{Vector of hierarchy levels.}
}
\value{
Vector of group names.
}
\description{
Translates hierarchy levels into unique group names which represent the
elements hierarchy position. An '_exp' is added at the end of the groups name
if it has child elements.
}

