% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treetable.R
\name{treetable}
\alias{treetable}
\title{Title}
\usage{
treetable(data, levels, css = NULL)
}
\arguments{
\item{data}{data.frame containing data to be shown}

\item{levels}{hierarchy levels}
}
\value{
html widget to display treetable data. See datatable function of
package DT
}
\description{
Title
}
\examples{
data <- data.frame(a = 1:9, x = c(9,8,7,6,5,4,3,2,1), b = c(5, 2, 3, 9, 4, 3, 1, 5, 3))
levels <- c(1, 2, 2, 1, 2, 3, 3, 2, 1)
css <- list(
  "thead" = list(
    "background-color" = "rgba(255,0,0,0.5)",
    color = "white"
  ),
  ".tree-level-1" = list(
    "background-color" = "rgba(255,0,0,0.5)"
  )
)
treetable(data, levels, css)
}

